// src/components/OverlayBubble/types.ts
export type bubbleType = {
  /**
   * Yellow bubble distance from top
   */
  s1top: number;

  /**
   * Green bubble distance from top
   */
  s2top: number;

  /**
   * Blue bubble distance from top
   */
  s3top: number;

  /**
   * Yellow bubble distance from left
   */
  s1left: number;

  /**
   * Green bubble distance from left
   */
  s2left: number;

  /**
   * Blue bubble distance from left
   */
  s3left: number;

  /**
   * The diamond shape distance from left
   */
  r1left: number;

  /**
   * The diamond shape distance from top
   */
  r1top: number;

  /**
   * The Rectangle shape distance from left
   */
  r2left: number;

  /**
   * The Rectangle shape distance from top
   */
  r2top: number;
}[];